api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
openai==1.12.0
python-multipart==0.0.9
pydantic==2.6.1
gunicorn==21.2.0 